refactor(require): relocate tests module
